Updates activity_sickness_entry.xml, strings.xml, AndroidManifest.xml, build.gradle, SicknessEntry.java, and SicknessEntryActivity.java files. User can now save an entry to the Firebase database with the correct lat/long information.

+++ app/src/main/java/com/a500/sweng/sickness_locator/models/SicknessEntry.java
@@ -11,8 +11,8 @@ public class SicknessEntry {
     public String sickness;
     public String severity;
     public int daysSick;
-    public float latitude;
-    public float longitude;
+    public double latitude;
+    public double longitude;
     public String userId;
 
     public SicknessEntry() {}
@@ -49,19 +49,19 @@ public class SicknessEntry {
         this.daysSick = daysSick;
     }
 
-    public float getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(float latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public float getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(float longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
